refactor: change location of map_tensor_dict
