+++ src/gt4sd/frameworks/torch/__init__.py
@@ -23,7 +23,7 @@
 #
 """Generic utils for pytorch."""
 
-from typing import List, Optional, Union
+from typing import Dict, List, Optional, Union
 
 import torch
 
@@ -95,3 +95,19 @@ def get_device_from_tensor(tensor: torch.Tensor) -> torch.device:
     device_id = tensor.get_device()
     device = "cpu" if device_id < 0 else f"cuda:{device_id}"
     return device_claim(device)
+
+
+def map_tensor_dict(
+    tensor_dict: Dict[str, torch.Tensor], device: torch.device
+) -> Dict[str, torch.Tensor]:
+    """
+    Maps a dictionary of tensors to a specific device.
+
+    Args:
+        tensor_dict: A dictionary of tensors.
+        device: The device to map the tensors to.
+
+    Returns:
+        A dictionary of tensors mapped to the device.
+    """
+    return {key: tensor.to(device) for key, tensor in tensor_dict.items()}
